+++ extensions/cornerstone/src/commandsModule.ts
@@ -25,7 +25,6 @@ import {
   colorPickerDialog,
 } from '@ohif/extension-default';
 import { vec3, mat4 } from 'gl-matrix';
-
 import CornerstoneViewportDownloadForm from './utils/CornerstoneViewportDownloadForm';
 import toggleImageSliceSync from './utils/imageSliceSync/toggleImageSliceSync';
 import { getFirstAnnotationSelected } from './utils/measurementServiceMappings/utils/selection';
@@ -34,6 +33,7 @@ import toggleVOISliceSync from './utils/toggleVOISliceSync';
 import { usePositionPresentationStore, useSegmentationPresentationStore } from './stores';
 import { toolNames } from './initCornerstoneTools';
 
+const { DefaultHistoryMemo } = csUtils.HistoryMemo;
 const toggleSyncFunctions = {
   imageSlice: toggleImageSliceSync,
   voi: toggleVOISliceSync,
@@ -140,19 +140,16 @@ function commandsModule({
       )?.[0];
 
       if (previousReferencedDisplaySetStoreKey) {
-        if (referencedImageId) {
-          setPositionPresentation(previousReferencedDisplaySetStoreKey, {
-            viewReference: {
-              referencedImageId,
-            },
-          });
-        } else {
-          setPositionPresentation(
-            previousReferencedDisplaySetStoreKey,
-            presentations.positionPresentation
-          );
-        }
+        const presentationData = referencedImageId
+          ? {
+              ...presentations.positionPresentation,
+              viewReference: {
+                referencedImageId,
+              },
+            }
+          : presentations.positionPresentation;
 
+        setPositionPresentation(previousReferencedDisplaySetStoreKey, presentationData);
         return;
       }
 
@@ -419,6 +416,37 @@ function commandsModule({
       }
 
       actions.setViewportWindowLevel({ ...props, viewportId });
+    },
+    setWindowLevelPreset: ({ presetName, presetIndex }) => {
+      const windowLevelPresets = customizationService.getCustomization(
+        'cornerstone.windowLevelPresets'
+      );
+
+      const activeViewport = viewportGridService.getActiveViewportId();
+      const viewport = cornerstoneViewportService.getCornerstoneViewport(activeViewport);
+      const metadata = viewport.getImageData().metadata;
+
+      const modality = metadata.Modality;
+
+      if (!modality) {
+        return;
+      }
+
+      const windowLevelPresetForModality = windowLevelPresets[modality];
+
+      if (!windowLevelPresetForModality) {
+        return;
+      }
+
+      const windowLevelPreset =
+        windowLevelPresetForModality[presetName] ??
+        Object.values(windowLevelPresetForModality)[presetIndex];
+
+      actions.setViewportWindowLevel({
+        viewportId: activeViewport,
+        window: windowLevelPreset.window,
+        level: windowLevelPreset.level,
+      });
     },
     setToolEnabled: ({ toolName, toggle, toolGroupId }) => {
       const { viewports } = viewportGridService.getState();
@@ -1317,6 +1345,12 @@ function commandsModule({
         measurementService.remove(activeAnnotationUID);
       });
     },
+    undo: () => {
+      DefaultHistoryMemo.undo();
+    },
+    redo: () => {
+      DefaultHistoryMemo.redo();
+    },
   };
 
   const definitions = {
@@ -1377,6 +1411,9 @@ function commandsModule({
     },
     setWindowLevel: {
       commandFn: actions.setWindowLevel,
+    },
+    setWindowLevelPreset: {
+      commandFn: actions.setWindowLevelPreset,
     },
     setToolActive: {
       commandFn: actions.setToolActive,
@@ -1577,6 +1614,8 @@ function commandsModule({
     deleteActiveAnnotation: {
       commandFn: actions.deleteActiveAnnotation,
     },
+    undo: actions.undo,
+    redo: actions.redo,
   };
 
   return {