+++ modes/microscopy/src/index.tsx
@@ -9,7 +9,7 @@ const ohif = {
   sopClassHandler: '@ohif/extension-default.sopClassHandlerModule.stack',
   hangingProtocols: '@ohif/extension-default.hangingProtocolModule.default',
   leftPanel: '@ohif/extension-default.panelModule.seriesList',
-  rightPanel: '@ohif/extension-default.panelModule.measure',
+  rightPanel: '@ohif/extension-dicom-microscopy.panelModule.measure',
 };
 
 export const cornerstone = {
@@ -90,7 +90,7 @@ function modeFactory({ modeConfiguration }) {
               leftPanelResizable: true,
               leftPanelClosed: true, // we have problem with rendering thumbnails for microscopy images
               rightPanelClosed: true, // we do not have the save microscopy measurements yet
-              rightPanels: ['@ohif/extension-dicom-microscopy.panelModule.measure'],
+              rightPanels: [ohif.rightPanel],
               rightPanelResizable: true,
               viewports: [
                 {
@@ -123,7 +123,6 @@ function modeFactory({ modeConfiguration }) {
       dicomvideo.sopClassHandler,
       dicompdf.sopClassHandler,
     ],
-    hotkeys: [...hotkeys.defaults.hotkeyBindings],
     ...modeConfiguration,
   };
 }