+++ modes/preclinical-4d/src/index.tsx
@@ -19,7 +19,6 @@ const ohif = {
   chartSopClassHandler: '@ohif/extension-default.sopClassHandlerModule.chart',
   hangingProtocol: '@ohif/extension-default.hangingProtocolModule.default',
   leftPanel: '@ohif/extension-default.panelModule.seriesList',
-  rightPanel: '@ohif/extension-default.panelModule.measure',
   chartViewport: '@ohif/extension-default.viewportModule.chartViewport',
 };
 
@@ -168,7 +167,6 @@ function modeFactory({ modeConfiguration }) {
     // general handler needs to come last.  For this case, the dicomvideo must
     // come first to remove video transfer syntax before ohif uses images
     sopClassHandlers: [ohif.chartSopClassHandler, ohif.defaultSopClassHandler],
-    hotkeys: [...hotkeys.defaults.hotkeyBindings],
   };
 }