+++ platform/core/src/defaults/index.js
@@ -1,4 +1,7 @@
 import hotkeyBindings from './hotkeyBindings';
 import windowLevelPresets from './windowLevelPresets';
-export { hotkeyBindings, windowLevelPresets };
-export default { hotkeyBindings, windowLevelPresets };
+
+export default {
+  hotkeyBindings,
+  windowLevelPresets,
+};